+++ user/apps/tests/syscall/gvisor/whitelist.txt
@@ -91,6 +91,8 @@ socket_unix_unbound_dgram_test
 socket_unix_unbound_filesystem_test
 socket_unix_unbound_stream_test
 raw_socket_test
+raw_socket_icmp_test
+raw_socket_hdrincl_test
 
 # 信号处理测试
 sigaction_test
